Move nullCriteriaRefMeasureObs to DetailedResultsHelpers file
